requirements.txt
PyMuPDF==1.20.2
Pillow==9.2.0
